Added "toComparisonCase" to TextUtils
For storing strings in a case-insensitive way, say, in a hash set, we
should transform them into a case-insensitive representation. This, so
far, was done by using toLowerCase. However, the default way to compare
strings in a case-insensitive manner in Java is using upper case (and,
in special situations, lower case, too). In order to remain more or less
independent of the current fashion, I decide to define a case-neutral
method name, toComparisonCase, which should be used in all methods
needing case-insensitive string storage. It currently maps to upper
casing. This may break some tests.