NetworkUtils: Replaced curlmyip.com with curlmyip.de
http://curlmyip.com/ seemingly no longer works, or at least I cannot
reach it from here. http://curlmyip.de seems to be a good alternative.

My NetworkUtils class has helper methods for obtaining the own global
IP address. This requires using some external tool to which we can send
a (HTTP) request and which print the IP address of the server from where
this request comes. Of course, this might be an NAT-ed address, i.e., it
is not clear whether we can actually reach our computer from somewhere
else. But, for now, this is the best we can do.

TODO: Open a server socket, then try to send a message to the IP address
obtained from the outside tools and check whether it arrives at the
server socket. That would be a good idea to verify the own IP.