Removed "validate()" from JobBuilder API, Clustering with Cluster Ranges
The validate() method has been removed from the job builder / tool API.
It seems to me that it actually is not really useful: Validation of data
should be done both by the setters of the job builders and the
constructor of the job objects. Ideally, checks would be placed into
package-private static methods.

The API for the clustering tools now allows you to specify a range of
clusters instead of only either a fixed number or nothing. The goal is
to give more freedom to the user.