Complete Redesign of "Execute" Class: No Longer ForkJoinPool Wrapper
So far, the class Execute was more or less a wrapper around
ForkJoinPool. However, this led to several problems, namely there were
scenarios where tasks recursively created other tasks and waited for
them in which I could not avoid getting RejecedExecutionExceptions.

The new Execute features a very simple pool of Threads that process work
from a joint queue. If one thread is supposed to wait until a given task
is done, it will no-blockingly check the queue for other work it can do
in the meantime. Thus all worker threads should be busy as much as
possible.