Added Optional null Checks to ArrayListView and ArraySetView Constructor
This forms another line of protection against null elements that should
not exist in lists. By providing a parameter deciding whether null is
permitted, we can optionally do an O(n) iteration over the list and
throw an IllegalArgumentException if null is encountered.

This led to several changes in other classes, which may now make use of
this option. This will cost a very small amount of additional runtime,
but may help to discover unexpected bugs.